requirements.txt
cassandra-driver~=3.28.0
langchain~=0.0.307
spacy~=3.7.0
spacy_langdetect~=0.1.2
spacy_download~=1.1.0
beautifulsoup4~=4.12.2
tiktoken~=0.5.1
openai~=0.28.1
cassio~=0.1.3
python-dotenv~=1.0.0
